Moving type to interface folder
